+++ src/Breadcrumbs/CacheHit.php
@@ -2,15 +2,9 @@
 
 namespace Honeybadger\HoneybadgerLaravel\Breadcrumbs;
 
-use Honeybadger\HoneybadgerLaravel\Facades\Honeybadger;
-use Illuminate\Cache\Events\CacheHit as LaravelCacheHit;
-
-class CacheHit extends Breadcrumb
+/**
+ * @deprecated Use {@link \Honeybadger\HoneybadgerLaravel\Events\CacheHit} instead
+ */
+class CacheHit extends \Honeybadger\HoneybadgerLaravel\Events\CacheHit
 {
-    public $handles = LaravelCacheHit::class;
-
-    public function handleEvent(LaravelCacheHit $event)
-    {
-        Honeybadger::addBreadcrumb('Cache hit', ['key' => $event->key], 'query');
-    }
 }