+++ src/Breadcrumbs/NotificationSending.php
@@ -2,16 +2,9 @@
 
 namespace Honeybadger\HoneybadgerLaravel\Breadcrumbs;
 
-use Honeybadger\HoneybadgerLaravel\Facades\Honeybadger;
-use Illuminate\Notifications\Events\NotificationSending as LaravelNotificationSending;
-
-class NotificationSending extends NotificationBreadcrumb
+/**
+ * @deprecated Use {@link \Honeybadger\HoneybadgerLaravel\Events\NotificationSending} instead
+ */
+class NotificationSending extends \Honeybadger\HoneybadgerLaravel\Events\NotificationSending
 {
-    public $handles = LaravelNotificationSending::class;
-
-    public function handleEvent(LaravelNotificationSending $event)
-    {
-        $metadata = parent::getMetadata($event);
-        Honeybadger::addBreadcrumb('Sending notification', $metadata, 'notification');
-    }
 }