+++ src/Breadcrumbs/NotificationSent.php
@@ -2,16 +2,9 @@
 
 namespace Honeybadger\HoneybadgerLaravel\Breadcrumbs;
 
-use Honeybadger\HoneybadgerLaravel\Facades\Honeybadger;
-use Illuminate\Notifications\Events\NotificationSent as LaravelNotificationSent;
-
-class NotificationSent extends NotificationBreadcrumb
+/**
+ * @deprecated Use {@link \Honeybadger\HoneybadgerLaravel\Events\NotificationSent} instead
+ */
+class NotificationSent extends \Honeybadger\HoneybadgerLaravel\Events\NotificationSent
 {
-    public $handles = LaravelNotificationSent::class;
-
-    public function handleEvent(LaravelNotificationSent $event)
-    {
-        $metadata = parent::getMetadata($event);
-        Honeybadger::addBreadcrumb('Notification sent', $metadata, 'notification');
-    }
 }